pyproject.toml
[project]
name = "lemonstation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "pynetworktables>=2021.0.0",
    "setuptools>=82.0.1",
    "wheel>=0.46.3",
]

[project.scripts]
lemon = "main:main"

[tool.setuptools]
py-modules = ["main"]
